moving encoder/decoder and AB to modules
