+++ mammoth/modules/__init__.py
@@ -14,8 +14,35 @@ from mammoth.modules.embeddings import Embeddings, PositionalEncoding
 from mammoth.modules.weight_norm import WeightNormConv2d
 from mammoth.modules.average_attn import AverageAttention
 from mammoth.modules.stable_embeddings import StableEmbedding
+from mammoth.modules.attention_bridge import AttentionBridge
+
+from mammoth.modules.encoder import EncoderBase
+from mammoth.modules.transformer_encoder import TransformerEncoder
+from mammoth.modules.mean_encoder import MeanEncoder
+
+from mammoth.modules.decoder import DecoderBase
+from mammoth.modules.transformer_decoder import TransformerDecoder
+
+
+str2enc = {
+    "transformer": TransformerEncoder,
+    "mean": MeanEncoder,
+}
+
+
+str2dec = {
+    "transformer": TransformerDecoder,
+}
+
 
 __all__ = [
+    "DecoderBase",
+    "TransformerDecoder",
+    "str2dec",
+    "EncoderBase",
+    "TransformerEncoder",
+    "MeanEncoder",
+    "str2enc",
     "Elementwise",
     "context_gate_factory",
     "ContextGate",
@@ -31,4 +58,5 @@ __all__ = [
     "AverageAttention",
     "CopyGeneratorLMLossCompute",
     "StableEmbedding",
+    "AttentionBridge",
 ]
